Add c2sTransactionSchemaRaw  along side c2sTransactionSchema
