+++ packages/schemas/src/transaction/index.ts
@@ -1,5 +1,11 @@
 import s2cTransactionSchemaRaw from './s2cRaw';
 import s2cTransactionSchema from './s2c';
+import c2sTransactionSchemaRaw from './c2sRaw';
 import c2sTransactionSchema from './c2s';
 
-export { s2cTransactionSchemaRaw, s2cTransactionSchema, c2sTransactionSchema };
+export {
+	s2cTransactionSchemaRaw,
+	s2cTransactionSchema,
+	c2sTransactionSchemaRaw,
+	c2sTransactionSchema,
+};
