+++ packages/schemas/src/transaction/s2c.ts
@@ -2,6 +2,7 @@
 import { resolveAccountType } from '../shared/utils/id';
 import s2cTransactionSchemaRaw from './s2cRaw';
 
+// This is only for the purpose of transforming the data. It is not used for refine methods
 const s2cTransactionSchema = s2cTransactionSchemaRaw.transform((data) => {
 	const { split } = data;
 	const isSettled = split?.splits.every((split) => split.debtStatus === 'PAID');
